split mona up into multiple files
