+++ package.json
@@ -2,7 +2,7 @@
   "name": "mona",
   "version": "0.8.2",
   "description": "Composable parsers",
-  "main": "index.js",
+  "main": "lib/index.js",
   "license": "CC0-1.0",
   "repository": {
     "type": "git",
@@ -22,18 +22,19 @@
   },
   "files": [
     "README.md",
-    "index.js"
+    "lib/*"
   ],
   "devDependencies": {
     "babel-cli": "^6.1.1",
     "babel-preset-es2015": "^6.0.15",
+    "documentation": "^3.0.3",
     "mocha": "^2.3.3",
     "standard": "^5.3.1"
   },
   "readmeFilename": "README.md",
   "scripts": {
-    "build": "babel --presets es2015 src/mona.js > index.js",
+    "build": "babel --presets es2015 src --out-dir lib",
     "prepublish": "npm run build",
-    "test": "standard src/mona*.js test/* && npm run build && mocha --reporter spec test/* test/**/*"
+    "test": "standard src/* test/* && npm run build && mocha --reporter spec test/* test/**/*"
   }
 }
